Continued refactoring the AlertBox
