+++ src/components/DropdownTask.jsx
@@ -1,20 +1,14 @@
 import React, { useEffect, useState } from 'react'
 import {
-    Menu, MenuItem, Button, ListItemIcon, ListItemText, Grid
+    Menu, MenuItem, Button, ListItemIcon, ListItemText, Grid,
 } from '@material-ui/core'
 import MoreHorizIcon from '@material-ui/icons/MoreHoriz'
 import { Delete, Edit } from '@material-ui/icons'
 import AlertBox from './AlertBox'
-import { DELETE_TASK } from '../graphql/task/taskQueries'
-import { TASKORDER } from '../graphql/fragments'
-import { useMutation, useApolloClient } from '@apollo/client'
 
 const DropdownTask = ({ columnId, taskId, handleEdit }) => {
     const [open, setOpen] = useState(false)
     const [anchorEl, setAnchorEl] = useState(null)
-    const alertMsg = "This action will permanently delete this task from the board and can't be later examined! Are you sure you want to delete it?"
-    const client = useApolloClient()
-    const [deleteTask] = useMutation(DELETE_TASK)
 
     const handleClick = (event) => {
         setAnchorEl(event.currentTarget)
@@ -24,39 +18,12 @@ const DropdownTask = ({ columnId, taskId, handleEdit }) => {
         setOpen(true)
         setAnchorEl(null)
     }
-    const deleteTaskById = () => {
-        deleteTask({
-            variables: {
-                taskId,
-            },
-        })
-    }
-
-    const deleteTaskFromCache = () => {
-        const idToBeDeleted = `Task:${taskId}`
-        const columnIdForCache = `Column:${columnId}`
-        const data = client.readFragment({
-            id: columnIdForCache,
-            fragment: TASKORDER,
-        })
-        const newTaskOrder = data.taskOrder.filter((id) => id !== taskId)
-
-        client.writeFragment({
-            id: columnIdForCache,
-            fragment: TASKORDER,
-            data: {
-                taskOrder: newTaskOrder,
-            },
-        })
-        client.cache.evict({ id: idToBeDeleted })
-    }
 
     useEffect(() => {
         if (handleEdit) {
             setAnchorEl(null)
         }
     }, [handleEdit])
-
 
     return (
         <Grid item>
@@ -93,10 +60,9 @@ const DropdownTask = ({ columnId, taskId, handleEdit }) => {
             <AlertBox
                 open={open}
                 setOpen={setOpen}
-                alertMsg={alertMsg}
-                deleteTaskById={deleteTaskById}
-                deleteTaskFromCache={deleteTaskFromCache}
-                type={'TASK'}
+                taskId={taskId}
+                columnId={columnId}
+                action="DELETE_TASK"
             />
         </Grid>
     )
